Experiments with Collection, JNode, Tile, etc
Honestly, the winner is still looking like the original Library and LibraryItem. The next step is going to be making Arrangement completely unrelated - it shouldn't be integrated with Library to begin with, like Focus. Arrangement should operate in reaction to LibraryItems dirty Toggle, and Focus reacts to Arrangements InProgress Toggle.